jffs2: kill garbage collect thread when filesystem is corrupted
Summary:

We have seen several instances of kernel crash caused by corrupted
jffs2 partition on CMM BMC. Instead of crash the system, let's
dump a warning and terminate jffs2 garbage collect thread: the BMC
will be still reachable in this case, and it's up to the system
administrator to decide how to fix the problem.

Test Plan:

1) booted up linux 5.15 on CMM and wedge successfully.

Reviewers: tfang, amithash, benwei, zhdaniel, mikechoi, #openbmc

Reviewed By: zhdaniel

Differential Revision: https://phabricator.intern.facebook.com/D16987396

@fb-note-start
This is a workaround but not official fix: the jffs2 upstream maintainer
was willing to help us but we cannot share the file system image as it
may contain Facebook confidential data. So we will keep the workaround
for now.
The workaround is needed for all BMC platforms depending on jffs2.
@fb-note-end