hwmon: max31790: support to config PWM as TACH
The PWM outputs of max31790 could be used as tachometer inputs by
setting the fan configuration register, but the driver doesn't support
to config the PWM outputs as tachometer inputs currently.

Add a function to get properties of the setting of max31790 to config
PWM outputs as tachometer inputs before initializing max31790.

This commit depends on the patch below:
[1] https://lore.kernel.org/all/20231102032834.3827289-1-Delphine_CC_Chiu@wiwynn.com/
[2] https://lore.kernel.org/all/20231102032834.3827289-2-Delphine_CC_Chiu@wiwynn.com/
[3] https://lore.kernel.org/all/20231102032834.3827289-3-Delphine_CC_Chiu@wiwynn.com/
[4] https://lore.kernel.org/all/20231102032834.3827289-4-Delphine_CC_Chiu@wiwynn.com/

Signed-off-by: Delphine CC Chiu <Delphine_CC_Chiu@wiwynn.com>