net: ncsi: check for netlink-driven responses before requiring a handler
Lore link:
https://lore.kernel.org/netdev/7346cdea74daedd0abecfcdcba2e7aa3be203a6e.camel@codeconstruct.com.au/T/#t

Currently, the NCSI response path will require an opcode-specific
handler for all incoming messages. However, we may be receiving a
response from a netlink-generated request, which may not have a
corresponding in-kernel handler for that request opcode.

Check for the pending request (and hence for NETLINK_DRIVEN) before
requiring a handler, deferring the check until we know it's a
kernel-driven command.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>